requirements.txt
bidict==0.21.2
click==8.0.1
cycler==0.10.0
dnspython==1.16.0
eventlet==0.31.0
Flask==2.0.1
Flask-MQTT==1.1.1
Flask-SocketIO==5.1.0
greenlet==1.1.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
npm==0.1.1
numpy==1.21.0
optional-django==0.1.0
paho-mqtt==1.5.1
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.2.0
python-socketio==5.3.0
simple-websocket==0.2.0
six==1.16.0
Werkzeug==2.0.1
wsproto==1.0.0
